+++ CHANGELOG.md
@@ -1,3 +1,80 @@
+## [101.4.7](https://github.com/dhis2/maps-app/compare/v101.4.6...v101.4.7) (2025-11-25)
+
+
+### Bug Fixes
+
+* make selected radio option visible in style tab [DHIS2-19809] ([#3589](https://github.com/dhis2/maps-app/issues/3589)) ([74e8655](https://github.com/dhis2/maps-app/commit/74e8655ac9e4367a1648d6fea55069d2ea0aed18))
+
+## [101.4.6](https://github.com/dhis2/maps-app/compare/v101.4.5...v101.4.6) (2025-11-05)
+
+
+### Bug Fixes
+
+* pin legend on click in dashboard [DHIS2-16467] ([#3576](https://github.com/dhis2/maps-app/issues/3576)) ([aa43735](https://github.com/dhis2/maps-app/commit/aa437350990d68e69c1b58544e5fede46a49f58a))
+
+## [101.4.5](https://github.com/dhis2/maps-app/compare/v101.4.4...v101.4.5) (2025-11-04)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([#3556](https://github.com/dhis2/maps-app/issues/3556)) ([d8982af](https://github.com/dhis2/maps-app/commit/d8982af1c0dbe09f25302e8557c49543f70eef33))
+
+## [101.4.4](https://github.com/dhis2/maps-app/compare/v101.4.3...v101.4.4) (2025-10-30)
+
+
+### Bug Fixes
+
+* use interpretations provider from analytics [DHIS2-19822] ([#3568](https://github.com/dhis2/maps-app/issues/3568)) ([96b9b5d](https://github.com/dhis2/maps-app/commit/96b9b5df9b43d207b82cdfbb7181bff216a2a300))
+
+## [101.4.3](https://github.com/dhis2/maps-app/compare/v101.4.2...v101.4.3) (2025-09-12)
+
+
+### Bug Fixes
+
+* calendar updates [DHIS2-19717] ([#3542](https://github.com/dhis2/maps-app/issues/3542)) ([9daf4de](https://github.com/dhis2/maps-app/commit/9daf4deaf740c27534f5371935cc0aa1b9fd437a))
+
+## [101.4.2](https://github.com/dhis2/maps-app/compare/v101.4.1...v101.4.2) (2025-08-19)
+
+
+### Bug Fixes
+
+* data table crashes on columns containing empty values [DHIS2-19267] ([#3564](https://github.com/dhis2/maps-app/issues/3564)) ([5152bb1](https://github.com/dhis2/maps-app/commit/5152bb12acd0aaa78662d647301b4a404369fe35))
+
+## [101.4.1](https://github.com/dhis2/maps-app/compare/v101.4.0...v101.4.1) (2025-08-13)
+
+
+### Bug Fixes
+
+* update cli app scripts to open links in same tab [DHIS2-19973] ([#3561](https://github.com/dhis2/maps-app/issues/3561)) ([4e91f02](https://github.com/dhis2/maps-app/commit/4e91f029dc953069f201fabf73edffbe32f6276b))
+
+# [101.4.0](https://github.com/dhis2/maps-app/compare/v101.3.2...v101.4.0) (2025-08-01)
+
+
+### Features
+
+* support DE/TEA of type OU and Event OU in Event layers [DHIS2-3354] ([#3511](https://github.com/dhis2/maps-app/issues/3511)) ([8240f5f](https://github.com/dhis2/maps-app/commit/8240f5fd2d03e64701288daf8638538cfda300d2))
+
+## [101.3.2](https://github.com/dhis2/maps-app/compare/v101.3.1...v101.3.2) (2025-07-30)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([#3528](https://github.com/dhis2/maps-app/issues/3528)) ([f25db38](https://github.com/dhis2/maps-app/commit/f25db38d896c9ddfada1ec54fae375c25467fd42))
+
+## [101.3.1](https://github.com/dhis2/maps-app/compare/v101.3.0...v101.3.1) (2025-07-09)
+
+
+### Bug Fixes
+
+* various bubble layer issues [DHIS2-15696] [DHIS2-19209] [DHIS2-19447] [DHIS2-19448] ([#3550](https://github.com/dhis2/maps-app/issues/3550)) ([8dbd273](https://github.com/dhis2/maps-app/commit/8dbd273d9d311bc6544284190c4262fdf1d719ae))
+
+# [101.3.0](https://github.com/dhis2/maps-app/compare/v101.2.15...v101.3.0) (2025-07-07)
+
+
+### Features
+
+* S2-EOX basemap and Azure as alternative to Bing [DHIS2-18500] [DHIS2-19152] ([#3543](https://github.com/dhis2/maps-app/issues/3543)) ([ebd109b](https://github.com/dhis2/maps-app/commit/ebd109bbbfbe9eef438e5ce7c11ec0387b1998e2))
+
 ## [101.2.15](https://github.com/dhis2/maps-app/compare/v101.2.14...v101.2.15) (2025-06-27)